+++ packages/terra-date-picker/src/DatePicker.jsx
@@ -88,13 +88,13 @@ const propTypes = {
    */
   onChangeRaw: PropTypes.func,
   /**
-   * **Deprecated**, A callback function to execute when clicking outside of the picker to dismiss it. Resolves to `onDismiss`.
+   * **Deprecated**, A callback function to execute when clicking outside of the picker to dismiss it. Resolves to `onRequestClose`.
    */
   onClickOutside: PropTypes.func,
   /**
-   * A callback function to execute when picker is dismissed. onDismiss(event)
+   * A callback function to execute when picker is dismissed. onRequestClose(event)
    */
-  onDismiss: PropTypes.func,
+  onRequestClose: PropTypes.func,
   /**
    * A callback function triggered when the date picker component receives focus.
    * This event does not get triggered when the focus is moved from the date input to the calendar button since the focus is still within the main date picker component.
@@ -398,7 +398,7 @@ class DatePicker extends React.Component {
       onChange,
       onChangeRaw,
       onClickOutside,
-      onDismiss,
+      onRequestClose,
       onFocus,
       onSelect,
       required,
@@ -459,7 +459,7 @@ class DatePicker extends React.Component {
             onChange={this.handleChange}
             onChangeRaw={this.handleChangeRaw}
             onClickOutside={this.handleOnClickOutside}
-            onDismiss={onDismiss}
+            onRequestClose={onRequestClose}
             onFocus={this.handleOnInputFocus}
             onSelect={this.handleOnSelect}
             required={required}