+++ packages/terra-date-picker/src/DatePickerField.jsx
@@ -100,13 +100,13 @@ const propTypes = {
    */
   onChangeRaw: PropTypes.func,
   /**
-   * **Deprecated**, A callback function to execute when clicking outside of the picker to dismiss it. Resolves to `onDismiss`.
+   * **Deprecated**, A callback function to execute when clicking outside of the picker to dismiss it. Resolves to `onRequestClose`.
    */
   onClickOutside: PropTypes.func,
   /**
-   * A callback function to execute when picker is dismissed. onDismiss(event)
+   * A callback function to execute when picker is dismissed. onRequestClose(event)
    */
-  onDismiss: PropTypes.func,
+  onRequestClose: PropTypes.func,
   /**
    * A callback function triggered when the date picker component receives focus.
    * This event does not get triggered when the focus is moved from the date input to the calendar button since the focus is still within the main date picker component.
@@ -198,7 +198,7 @@ const DatePickerField = (props) => {
     onChange,
     onChangeRaw,
     onClickOutside,
-    onDismiss,
+    onRequestClose,
     onFocus,
     onSelect,
     required,
@@ -259,7 +259,7 @@ const DatePickerField = (props) => {
         onBlur={onBlur}
         onChange={onChange}
         onChangeRaw={onChangeRaw}
-        onDismiss={onDismiss}
+        onRequestClose={onRequestClose}
         onClickOutside={onClickOutside}
         onFocus={onFocus}
         onSelect={onSelect}